Revert "Re-add debugging code"
This reverts commit e6228e384b9ed2bd3e2d481df89f517ace53e66c.

Effectively this simplifies the code by ignoring our async code
standards for the ratevalidator package. It was too complicated to
adhere to the standards and ratevalidator does not actually need to do
any cleanup (for now) when the context is canceled. This also removes
some changes that were only made for debugging purposes.